ACCUMULO-1973 Add a warning if we don't find non-default values for namenode configuration
This is one of biggest confusions that new users seem to run into when
trying to initialize Accumulo (as it's a problem with us not finding
Hadoop configurations). Hadoop will work properly, as its configuration
is likely correct, but Accumulo can't find core-site.xml, tries to write
to the root of the local fs, and fails. Try to be a bit more explicit
about what happened when we initialized and the Configuration from the FS.

Reason: Polish
Author: Josh Elser <elserj@apache.org>
Ref: CDH-16082